remove mockcontents as separate package
