--- packages/tree-finder-mockcontents/src/index.ts
@@ -1,7 +1,0 @@
-/*
- * Copyright (c) 2020, Max Klein
- *
- * This file is part of the tree-finder library, distributed under the terms of
- * the BSD 3 Clause license. The full license can be found in the LICENSE file.
- */
-export * from "./mockcontent";
