--- packages/tree-finder-mockcontents/tsconfig.json
@@ -1,16 +1,0 @@
-{
-    "$schema": "http://json.schemastore.org/tsconfig",
-    "extends": "../../tsconfigbase",
-    "compilerOptions": {
-        // "baseUrl": ".",
-        "declarationDir": "types",
-        "outDir": "lib",
-        "rootDir": "src"
-    },
-    "include": ["src/**/*"],
-    "references": [
-        {
-            "path": "../tree-finder"
-        }
-    ]
-}
